Avoid potential int overflow

+++ processor/tailsamplingprocessor/internal/sampling/span_count_sampler.go
@@ -15,8 +15,8 @@ import (
 
 type spanCount struct {
 	logger   *zap.Logger
-	minSpans int32
-	maxSpans int32
+	minSpans int64
+	maxSpans int64
 }
 
 var _ samplingpolicy.Evaluator = (*spanCount)(nil)
@@ -25,8 +25,8 @@ var _ samplingpolicy.Evaluator = (*spanCount)(nil)
 func NewSpanCount(settings component.TelemetrySettings, minSpans, maxSpans int32) samplingpolicy.Evaluator {
 	return &spanCount{
 		logger:   settings.Logger,
-		minSpans: minSpans,
-		maxSpans: maxSpans,
+		minSpans: int64(minSpans),
+		maxSpans: int64(maxSpans),
 	}
 }
 
@@ -34,7 +34,7 @@ func NewSpanCount(settings component.TelemetrySettings, minSpans, maxSpans int32
 func (c *spanCount) Evaluate(_ context.Context, _ pcommon.TraceID, traceData *samplingpolicy.TraceData) (samplingpolicy.Decision, error) {
 	c.logger.Debug("Evaluating spans counts in filter")
 
-	spanCount := int32(traceData.SpanCount)
+	spanCount := traceData.SpanCount
 	switch {
 	case c.maxSpans == 0 && spanCount >= c.minSpans:
 		return samplingpolicy.Sampled, nil
